app/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
psycopg2
python-multipart
python-jose[cryptography]
passlib[bcrypt]
firebase-admin==5.2.0
Unidecode==1.3.4